v1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo==4.6.1
pydantic==2.5.2
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sliver-python==1.5.37
python-dotenv==1.0.0
email-validator==2.1.0.post1
geoip2==4.7.0
bcrypt==4.0.1 